pyproject.toml
[project]
name = "mapillary-downloader"
version = "0.1.0"
description = "Download street-level images from Mapillary for a given city"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "httpx>=0.28.1",
    "mercantile>=1.2.1",
    "python-dotenv>=1.2.1",
]

[project.scripts]
mapillary-download = "mapillary_downloader.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mapillary_downloader"]
